+++ lib/persistence.dart
@@ -17,4 +17,5 @@ export 'src/persistence/types/point.dart';
 
 export 'src/persistence/database_adapter.dart';
 export 'src/persistence/database_connection.dart';
+export 'src/persistence/database_context.dart';
 export 'src/persistence/persistence_exception.dart';