+++ lib/src/persistence/database_connection.dart
@@ -1,5 +1,7 @@
 import 'package:datahub/datahub.dart';
 
+import 'database_context.dart';
+
 /// Represents a single connection to a database.
 ///
 /// A [DatabaseConnection] is acquired by using a [DatabaseAdapter].
@@ -22,46 +24,14 @@ abstract class DatabaseConnection {
   /// be used anymore.
   Future<void> close();
 
-  Future<List<TDao>> query<TDao>(
-    DaoDataBean<TDao> bean, {
-    Filter filter = Filter.empty,
-    Sort sort = Sort.empty,
-    int offset = 0,
-    int limit = -1,
-  });
-
-  Future<TDao?> queryId<TDao, TPrimaryKey>(
-      PKDaoDataBean<TDao, TPrimaryKey> bean, TPrimaryKey id);
-
-  Future<bool> idExists<TPrimaryKey>(
-      PrimaryKeyDataBean<TPrimaryKey> bean, TPrimaryKey id);
-
-  Future<List<dynamic>> select(
-    QuerySource bean,
-    List<QuerySelect> select, {
-    Filter filter = Filter.empty,
-    Sort sort = Sort.empty,
-    int offset = 0,
-    int limit = -1,
-  });
-
-  /// Returns primary key of inserted object.
-  Future<dynamic> insert<TDao extends BaseDao>(TDao object);
-
-  Future<void> update<TDao extends PKBaseDao>(TDao object);
-
-  Future<void> updateId<TPrimaryKey>(PrimaryKeyDataBean<TPrimaryKey> bean,
-      TPrimaryKey id, Map<String, dynamic> values);
-
-  /// Returns number of affected rows.
-  Future<int> updateWhere(
-      BaseDataBean bean, Map<String, dynamic> values, Filter filter);
-
-  Future<void> delete<TDao extends PKBaseDao>(TDao object);
-
-  Future<void> deleteId<TPrimaryKey>(
-      PrimaryKeyDataBean<TPrimaryKey> bean, TPrimaryKey id);
-
-  /// Returns number of affected rows.
-  Future<int> deleteWhere(DaoDataBean bean, Filter filter);
+  /// Runs [delegate] inside a transaction.
+  ///
+  /// If [delegate] returns without throwing an exception, the transaction
+  /// is committed and the return value is passed through as return value
+  /// of this method.
+  ///
+  /// If [delegate] throws an exception, the transaction is rolled back
+  /// and the exception is rethrown.
+  Future<T> runTransaction<T>(
+      Future<T> Function(DatabaseContext context) delegate);
 }