+++ lib/src/persistence/postgresql/postgresql_database_adapter.dart
@@ -1,4 +1,5 @@
 import 'package:datahub/datahub.dart';
+import 'package:datahub/src/persistence/postgresql/postgresql_database_context.dart';
 
 // ignore: import_of_legacy_library_into_null_safe
 import 'package:postgres/postgres.dart' as postgres;
@@ -65,58 +66,61 @@ class PostgreSQLDatabaseAdapter extends DatabaseAdapter {
   Future<void> initializeSchema({bool ignoreMigration = false}) async {
     final connection = await _connection();
 
-    if (await _schemaExists(connection)) {
-      if (!ignoreMigration) {
-        String? versionString;
-        try {
-          versionString = await connection.getMetaValue(schemaVersionKey);
-        } catch (e) {
-          throw PersistenceException(
-            'Could not query version for schema "${schema.name}".',
-            cause: e,
-          );
+    await connection.runTransaction((abstractContext) async {
+      final context = abstractContext as PostgreSQLDatabaseContext;
+
+      if (await _schemaExists(context)) {
+        if (!ignoreMigration) {
+          String? versionString;
+          try {
+            versionString = await context.getMetaValue(schemaVersionKey);
+          } catch (e) {
+            throw PersistenceException(
+              'Could not query version for schema "${schema.name}".',
+              cause: e,
+            );
+          }
+
+          if (versionString == null) {
+            throw PersistenceException(
+                'Schema "${schema.name}" does not provide version.');
+          }
+
+          final version = int.parse(versionString);
+          if (version != schema.version) {
+            resolve<LogService>().i(
+                'Migrating database schema from v$version to v${schema.version}.',
+                sender: 'DataHub');
+            final migrator = PostgreSQLDatabaseMigrator(schema, context);
+
+            await schema.migrate(migrator, version);
+            await context.setMetaValue(
+                schemaVersionKey, schema.version.toString());
+          }
         }
+      } else {
+        await context.execute(RawSql('CREATE SCHEMA ${schema.name}'));
 
-        if (versionString == null) {
-          throw PersistenceException(
-              'Schema "${schema.name}" does not provide version.');
-        }
+        await context.execute(CreateTableBuilder(schema.name, metaTable)
+          ..fields.addAll([
+            PrimaryKey(FieldType.String, metaTable, 'key'),
+            DataField(FieldType.String, metaTable, 'value')
+          ]));
 
-        final version = int.parse(versionString);
-        if (version != schema.version) {
-          resolve<LogService>().i(
-              'Migrating database schema from v$version to v${schema.version}.',
-              sender: 'DataHub');
-          final migrator = PostgreSQLDatabaseMigrator(schema, connection);
+        await context.setMetaValue(schemaVersionKey, schema.version.toString());
 
-          await schema.migrate(migrator, version);
-          await connection.setMetaValue(
-              schemaVersionKey, schema.version.toString());
+        // create scheme
+        for (final layout in schema.beans) {
+          await context.execute(CreateTableBuilder.fromLayout(schema, layout));
         }
       }
-    } else {
-      await connection.execute(RawSql('CREATE SCHEMA ${schema.name}'));
-
-      await connection.execute(CreateTableBuilder(schema.name, metaTable)
-        ..fields.addAll([
-          PrimaryKey(FieldType.String, metaTable, 'key'),
-          DataField(FieldType.String, metaTable, 'value')
-        ]));
-
-      await connection.setMetaValue(
-          schemaVersionKey, schema.version.toString());
-
-      // create scheme
-      for (final layout in schema.beans) {
-        await connection.execute(CreateTableBuilder.fromLayout(schema, layout));
-      }
-    }
+    });
 
     _initialized = true;
   }
 
-  Future<bool> _schemaExists(PostgreSQLDatabaseConnection connection) async {
-    final result = await connection.querySql(RawSql(
+  Future<bool> _schemaExists(PostgreSQLDatabaseContext context) async {
+    final result = await context.querySql(RawSql(
         'SELECT schema_name FROM information_schema.schemata '
         'WHERE schema_name = @name;',
         {'name': schema.name}));