+++ lib/src/persistence/postgresql/postgresql_database_context.dart
@@ -1,0 +1,203 @@
+import 'package:boost/boost.dart';
+import 'package:datahub/datahub.dart';
+import 'package:postgres/postgres.dart' as postgres;
+
+import 'sql/sql.dart';
+
+class PostgreSQLDatabaseContext implements DatabaseContext {
+  static const _metaKeyColumn = 'key';
+  static const _metaValueColumn = 'value';
+  final PostgreSQLDatabaseAdapter _adapter;
+  final postgres.PostgreSQLExecutionContext _context;
+
+  PostgreSQLDatabaseContext(this._adapter, this._context);
+
+  Future<String?> getMetaValue(String key) async {
+    final result = await _context.query(
+        'SELECT "value" FROM ${_adapter.schema.name}.$metaTable WHERE "key" = @key',
+        substitutionValues: {'key': key});
+
+    if (result.isNotEmpty) {
+      return result.firstOrNull?.firstOrNull as String?;
+    } else {
+      return null;
+    }
+  }
+
+  Future<void> setMetaValue(String key, String value) async {
+    final currentValue = await getMetaValue(key);
+    if (currentValue == null) {
+      await _context.execute(
+          'INSERT INTO ${_adapter.schema.name}.$metaTable ("$_metaKeyColumn", "$_metaValueColumn") VALUES (@key, @value)',
+          substitutionValues: {'key': key, 'value': value});
+    } else {
+      await _context.execute(
+          'UPDATE ONLY ${_adapter.schema.name}.$metaTable SET "$_metaValueColumn" = @value WHERE "$_metaKeyColumn" = @key',
+          substitutionValues: {'key': key, 'value': value});
+    }
+  }
+
+  Future<int> execute(SqlBuilder builder) async {
+    final result = builder.buildSql();
+    return await _context.execute(result.a, substitutionValues: result.b);
+  }
+
+  Future<List<Map<String, dynamic>>> querySql(SqlBuilder builder) async {
+    final builderResult = builder.buildSql();
+    final result = await _context.query(builderResult.a,
+        substitutionValues: builderResult.b);
+    return result
+        .map((row) => row
+            .toColumnMap()
+            .map((key, value) => MapEntry(key, _fromSqlData(value))))
+        .toList();
+  }
+
+  @override
+  Future<List<TDao>> query<TDao>(
+    DaoDataBean<TDao> bean, {
+    Filter filter = Filter.empty,
+    Sort sort = Sort.empty,
+    int offset = 0,
+    int limit = -1,
+  }) async {
+    final from = SelectFromTable(_adapter.schema.name, bean.layoutName);
+    final result = await querySql(SelectBuilder(from)
+      ..select([const WildcardSelect()])
+      ..where(filter)
+      ..orderBy(sort)
+      ..offset(offset)
+      ..limit(limit));
+    return result.map(bean.map).toList();
+  }
+
+  @override
+  Future<TDao?> queryId<TDao, TPrimaryKey>(
+      PKDaoDataBean<TDao, TPrimaryKey> bean, TPrimaryKey id) async {
+    final primaryKey = bean.primaryKeyField;
+
+    final from = SelectFromTable(_adapter.schema.name, bean.layoutName);
+    final result = await querySql(
+        SelectBuilder(from)..where(Filter.equals(primaryKey, id)));
+
+    return result.map(bean.map).firstOrNull;
+  }
+
+  @override
+  Future<bool> idExists<TPrimaryKey>(
+      PrimaryKeyDataBean<TPrimaryKey> bean, TPrimaryKey id) async {
+    final primaryKey = bean.primaryKeyField;
+
+    final from = SelectFromTable(_adapter.schema.name, bean.layoutName);
+    final result = await querySql(SelectBuilder(from)
+      ..select([primaryKey])
+      ..where(Filter.equals(primaryKey, id)));
+
+    return result.isNotEmpty;
+  }
+
+  @override
+  Future<dynamic> insert<TDao extends BaseDao>(TDao entry) async {
+    final bean = entry.bean;
+
+    final primaryKeyField = bean is PrimaryKeyDataBean
+        ? (bean as PrimaryKeyDataBean).primaryKeyField
+        : null;
+
+    final returning = primaryKeyField != null
+        ? SqlBuilder.escapeName(primaryKeyField.name)
+        : null;
+
+    final withPrimary = !(primaryKeyField?.autoIncrement ?? false);
+
+    final data = bean.unmap(entry, includePrimaryKey: withPrimary);
+    final result = await querySql(
+        InsertBuilder(_adapter.schema.name, entry.bean.layoutName)
+          ..values(data)
+          ..returning(returning));
+
+    return result.firstOrNull?.values.firstOrNull;
+  }
+
+  @override
+  Future<void> update<TDao extends PKBaseDao>(TDao object) async {
+    final bean = object.bean;
+    final data = bean.unmap(object);
+
+    final from = SelectFromTable(_adapter.schema.name, bean.layoutName);
+    await execute(UpdateBuilder(from)
+      ..values(data)
+      ..where(_pkFilter(bean, object.getPrimaryKey())));
+  }
+
+  @override
+  Future<void> updateId<TPrimaryKey>(PrimaryKeyDataBean<TPrimaryKey> bean,
+      TPrimaryKey id, Map<String, dynamic> values) async {
+    final from = SelectFromTable(_adapter.schema.name, bean.layoutName);
+    await execute(UpdateBuilder(from)
+      ..values(values)
+      ..where(_pkFilter(bean, id)));
+  }
+
+  @override
+  Future<int> updateWhere(
+      BaseDataBean bean, Map<String, dynamic> values, Filter filter) async {
+    final from = SelectFromTable(_adapter.schema.name, bean.layoutName);
+    return await execute(UpdateBuilder(from)
+      ..values(values)
+      ..where(filter));
+  }
+
+  @override
+  Future<void> delete<TDao extends PKBaseDao>(TDao object) async {
+    final bean = object.bean;
+    final from = SelectFromTable(_adapter.schema.name, bean.layoutName);
+    await execute(
+        DeleteBuilder(from)..where(_pkFilter(bean, object.getPrimaryKey())));
+  }
+
+  @override
+  Future<void> deleteId<TPrimaryKey>(
+      PrimaryKeyDataBean<TPrimaryKey> bean, dynamic id) async {
+    final from = SelectFromTable(_adapter.schema.name, bean.layoutName);
+    await execute(DeleteBuilder(from)..where(_pkFilter(bean, id)));
+  }
+
+  @override
+  Future<int> deleteWhere(BaseDataBean bean, Filter filter) async {
+    final from = SelectFromTable(_adapter.schema.name, bean.layoutName);
+    return await execute(DeleteBuilder(from)..where(filter));
+  }
+
+  @override
+  Future<List> select(
+    QuerySource source,
+    List<QuerySelect> select, {
+    Filter filter = Filter.empty,
+    Sort sort = Sort.empty,
+    int offset = 0,
+    int limit = -1,
+  }) {
+    final from = SelectFrom.fromQuerySource(_adapter.schema.name, source);
+    return querySql(SelectBuilder(from)
+      ..where(filter)
+      ..orderBy(sort)
+      ..offset(offset)
+      ..limit(limit)
+      ..select(select));
+  }
+
+  dynamic _fromSqlData(dynamic value) {
+    if (value is postgres.PgPoint) {
+      return Point(value.latitude, value.longitude);
+    }
+
+    return value;
+  }
+
+  Filter _pkFilter<TPrimaryKey>(
+      PrimaryKeyDataBean<TPrimaryKey> layout, TPrimaryKey id) {
+    final primaryKey = layout.primaryKeyField;
+    return Filter.equals(primaryKey, id);
+  }
+}